requirements.txt
fastapi
uvicorn
azure-cosmos
langchain
openai
python-dotenv
streamlit
